refactor(room): extract remaining executors, shrink useActionOrchestrator (C10)
Extract all remaining 6 switch cases (actionPrompt, confirmTrigger,
wolfRobotViewHunterStatus, multiSelectToggle, multiSelectConfirm,
groupConfirmAck) into dedicated executor files. Remove the legacy
switch entirely; useActionOrchestrator is now a registry dispatch
shell at 338 lines (down from 785).

New files:
- promptExecutor.ts (actionPrompt + confirmTrigger)
- wolfRobotExecutor.ts (wolfRobotViewHunterStatus)
- multiSelectExecutor.ts (multiSelectToggle + multiSelectConfirm)
- groupConfirmExecutor.ts (groupConfirmAck)

Also removed proceedWithActionTyped/ActionExtra/WitchStepResultsExtra
(no longer needed after full executor migration), simplified
confirmThenAct to use proceedWithAction directly.

Updated contract tests to reference executor files.
