+++ src/screens/RoomScreen/executors/index.ts
@@ -19,6 +19,19 @@ registerExecutor('wolfVote', wolfVoteExecutor);
 registerExecutor('actionConfirm', actionConfirmExecutor);
 registerExecutor('skip', skipExecutor);
 
+// ── C10 executors ──────────────────────────────────────────────────────────
+import { groupConfirmAckExecutor } from './groupConfirmExecutor';
+import { multiSelectConfirmExecutor, multiSelectToggleExecutor } from './multiSelectExecutor';
+import { actionPromptExecutor, confirmTriggerExecutor } from './promptExecutor';
+import { wolfRobotViewHunterStatusExecutor } from './wolfRobotExecutor';
+
+registerExecutor('actionPrompt', actionPromptExecutor);
+registerExecutor('confirmTrigger', confirmTriggerExecutor);
+registerExecutor('wolfRobotViewHunterStatus', wolfRobotViewHunterStatusExecutor);
+registerExecutor('multiSelectToggle', multiSelectToggleExecutor);
+registerExecutor('multiSelectConfirm', multiSelectConfirmExecutor);
+registerExecutor('groupConfirmAck', groupConfirmAckExecutor);
+
 // ── Public API ─────────────────────────────────────────────────────────────
 export { dispatchIntent } from './registry';
 export type { ExecutorContext, ExecutorMap, IntentExecutor } from './types';
