+++ src/screens/RoomScreen/executors/wolfRobotExecutor.ts
@@ -1,0 +1,71 @@
+/**
+ * wolfRobotExecutor — Handles 'wolfRobotViewHunterStatus' ActionIntent
+ *
+ * Shows hunter shoot-status dialog for wolfRobot disguise phase,
+ * then sends status-viewed acknowledgment via sendWolfRobotHunterStatusViewed.
+ */
+
+import { handleError } from '@/utils/errorPipeline';
+import { roomScreenLog } from '@/utils/logger';
+
+import type { IntentExecutor } from './types';
+
+export const wolfRobotViewHunterStatusExecutor: IntentExecutor = (_intent, ctx) => {
+  const {
+    gameState,
+    effectiveSeat,
+    currentSchema,
+    pendingHunterStatusViewed,
+    setPendingHunterStatusViewed,
+    sendWolfRobotHunterStatusViewed,
+    actionDialogs,
+  } = ctx;
+
+  if (!gameState?.wolfRobotReveal) return;
+
+  if (pendingHunterStatusViewed) {
+    roomScreenLog.debug('[wolfRobotViewHunterStatus] Skipping - pending submission');
+    return;
+  }
+
+  if (currentSchema?.id !== 'wolfRobotLearn') {
+    throw new Error(
+      `[RoomScreen] wolfRobotViewHunterStatus intent received but currentSchema is ${currentSchema?.id}, expected wolfRobotLearn`,
+    );
+  }
+
+  const dialogTitle = currentSchema.ui?.hunterGateDialogTitle;
+  const canShootText = currentSchema.ui?.hunterGateCanShootText;
+  const cannotShootText = currentSchema.ui?.hunterGateCannotShootText;
+
+  if (!dialogTitle || !canShootText || !cannotShootText) {
+    throw new Error(
+      '[RoomScreen] wolfRobotLearn schema missing hunterGate UI fields - schema-driven UI requires these',
+    );
+  }
+
+  const canShoot = gameState.wolfRobotReveal.canShootAsHunter === true;
+  const statusMessage = canShoot ? canShootText : cannotShootText;
+
+  actionDialogs.showRoleActionPrompt(dialogTitle, statusMessage, async () => {
+    if (effectiveSeat === null) {
+      roomScreenLog.warn(
+        '[wolfRobotViewHunterStatus] Cannot submit without seat (effectiveSeat is null)',
+      );
+      return;
+    }
+    setPendingHunterStatusViewed(true);
+    try {
+      await sendWolfRobotHunterStatusViewed(effectiveSeat);
+    } catch (error) {
+      handleError(error, {
+        label: '狼人机器人确认犵人状态',
+        logger: roomScreenLog,
+        alertTitle: '确认失败',
+        alertMessage: '状态确认发送失败，请稍后重试。如问题持续，请联系房主。',
+      });
+    } finally {
+      setPendingHunterStatusViewed(false);
+    }
+  });
+};
